requirements.txt
fastapi
pydantic
uvicorn
groq
python-dotenv
faker
